* make two free functions private to internal package & test w/ whitebox tests * Fix tests that failed in CI b/c assumed .filecoin dir exists.
